rename: api폴더 공통화
